--- a/benchmark/python/benchmark_hail/run/methods_benchmarks.py
+++ b/benchmark/python/benchmark_hail/run/methods_benchmarks.py
@@ -99,9 +99,9 @@ def pc_relate(mt_path):
     rel._force_count()
 
 
-@benchmark(args=profile_25.handle('mt'))
-def pc_relate_big(mt_path):
-    mt = hl.balding_nichols_model(3, 2 * 4096, 2 * 4096).checkpoint('/tmp/foo.mt', overwrite=True)
+@benchmark()
+def pc_relate_big():
+    mt = hl.balding_nichols_model(3, 2 * 4096, 2 * 4096)
     mt = mt.annotate_cols(scores = hl.range(2).map(lambda x: hl.rand_unif(0, 1)))
     rel = hl.pc_relate(mt.GT,
                        0.05,