--- a/benchmark/python/benchmark_hail/run/methods_benchmarks.py
+++ b/benchmark/python/benchmark_hail/run/methods_benchmarks.py
@@ -101,7 +101,7 @@ def pc_relate(mt_path):
 
 @benchmark()
 def pc_relate_big():
-    mt = hl.balding_nichols_model(3, 2 * 4096, 2 * 4096)
+    mt = hl.balding_nichols_model(3, 2 * 4096, 2 * 4096).checkpoint(hl.utils.new_temp_file(suffix='mt'))
     mt = mt.annotate_cols(scores = hl.range(2).map(lambda x: hl.rand_unif(0, 1)))
     rel = hl.pc_relate(mt.GT,
                        0.05,